+++ fiddles/bash/fiddle-0016-ForLoops/script.sh
@@ -1,17 +1,31 @@
-#!/usr/bin/env bash
+#!/bin/bash
 
 clear;
 echo "$0" | sed 's/\.\///g' | awk '{print toupper($0)}'
+echo "Bash version ${BASH_VERSION}..."
 
-function controlFunction() {
-    echo " hello $1"
+function exampleA() {
+    echo ""
+    echo "exampleA:"
+    echo ""
+    for i in 1 2 3 4 5
+    do
+       echo -e "\tWelcome $i times"
+    done
+
 }
 
-function valueFunction() {
-    z=`expr $1 + 3`;
-    echo ${z};
+function exampleB() {
+    echo ""
+    echo "exampleB:"
+    echo ""
+    for i in {1..5}
+    do
+        echo -e "\tWelcome $i times"
+    done
+    echo ""
 }
-controlFunction;
-controlFunction $(whoami;);
-rc=$(valueFunction 2;);
-echo "2 + 3 = ${rc}";
+
+
+exampleA
+exampleB