+++ fiddles/bash/template/script.sh
@@ -2,6 +2,7 @@
 
 clear;
 echo "$0" | sed 's/\.\///g' | awk '{print toupper($0)}'
+echo "Bash version ${BASH_VERSION}..."
 
 function controlFunction() {
     echo " hello $1"