+++ fiddles/extjs6/fiddle-0001-ItDoesntWork/app.js
@@ -1,0 +1,60 @@
+var meta = {
+    fiddleHeader: 'Version Info',
+    fiddleSubHeader: 'ExtJS 6 ~ version info.' + '<br />'
+};
+
+Ext.define('Fiddle.View', {
+    extend: 'Ext.panel.Panel',
+    html: 'Ext version ~ ' + Ext.getVersion().version
+});
+// Boiler plate
+Ext.define('App.BoxModel', {
+    extend: 'Ext.app.ViewModel',
+    alias: 'viewmodel.box',
+    data: {
+        header: meta.fiddleHeader,
+        subheader: meta.fiddleSubHeader
+    }
+});
+Ext.define('App.Box', {
+    extend: "Ext.container.Container",
+    border: true,
+    padding: 25,
+    viewModel: {
+        type: 'box'
+    },
+    items: [{
+        xtype: 'panel',
+        bind: {
+            title: '{header}'
+        },
+        items: [{
+            xtype: 'panel',
+            padding: 10,
+            border: false,
+            bind: {
+                html: '{subheader}'
+            }
+        }],
+        region: 'north'
+    }]
+});
+Ext.onReady(function() {
+    var fiddle = Ext.create('Fiddle.View'),
+        win = Ext.create('Ext.Window', {
+            title: meta.fiddleHeader,
+            closable: false,
+            height: 100,
+            width: 200,
+            layout: 'fit',
+            items: fiddle
+        }),
+        positionX = 25,
+        positionY = 192;
+    win.showAt([positionX, positionY]);
+    Ext.QuickTips.init();
+
+    Ext.create('App.Box', {
+        renderTo: Ext.getBody()
+    });
+});