+++ fiddles/extjs6/template/app.js
@@ -1,0 +1,96 @@
+var meta = {
+    fiddleHeader: '{{FiddleName}}',
+    fiddleSubHeader: 'Template Fiddle created @ {{BornOnDate}}' +
+    '<br />'
+};
+Ext.define('Fiddle.Records', {
+    extend: 'Ext.data.Store',
+    config: {
+        autoDestroy: true,
+        proxy: {
+            type: 'ajax',
+            url: 'data.json',
+            reader: {
+                type: 'json',
+                rootProperty: 'data',
+                idProperty: '_id',
+                totalProperty: 'total'
+            }
+        },
+        remoteSort: false,
+        sortInfo: {
+            field: 'name',
+            direction: 'DESC'
+        },
+        pageSize: 50,
+        fields: ['_id', 'index', 'isActive', 'checkingBalance', 'savingBalance', 'picture', 'age', 'eyeColor', 'name', 'gender', 'company', 'email', 'address', 'about', 'latitude', 'longitude'],
+        autoLoad: true
+    }
+}, function () {
+    Ext.create('Fiddle.Records', {
+        storeId: 'records'
+    });
+    Ext.create('Ext.data.ChainedStore', {
+        source: 'records',
+        storeId: 'viewrecords'
+    });
+});
+Ext.define('Fiddle.View', {
+    extend: 'Ext.panel.Panel',
+    html: 'Ext version ~ ' + Ext.getVersion().version
+});
+// Boiler plate
+Ext.define('App.BoxModel', {
+    extend: 'Ext.app.ViewModel',
+    alias: 'viewmodel.box',
+    data: {
+        header: meta.fiddleHeader,
+        subheader: meta.fiddleSubHeader
+    }
+});
+Ext.define('App.Box', {
+    extend: "Ext.container.Container",
+    border: true,
+    padding: 25,
+    viewModel: {
+        type: 'box'
+    },
+    items: [
+        {
+            xtype: 'panel',
+            bind: {
+                title: '{header}'
+            },
+            items: [
+                {
+                    xtype: 'panel',
+                    padding: 10,
+                    border: false,
+                    bind: {
+                        html: '{subheader}'
+                    }
+                }
+            ],
+            region: 'north'
+        }
+    ]
+});
+Ext.onReady(function () {
+    var fiddle = Ext.create('Fiddle.View'),
+        win = Ext.create('Ext.Window', {
+            title: meta.fiddleHeader,
+            closable: false,
+            height: 500,
+            width: 700,
+            layout: 'fit',
+            items: fiddle
+        }),
+        positionX = 25,
+        positionY = 192;
+    win.showAt([positionX, positionY]);
+    Ext.QuickTips.init();
+
+    Ext.create('App.Box', {
+        renderTo: Ext.getBody()
+    });
+});