+++ scripts/fiddle-index.sh
@@ -15,6 +15,7 @@
 # 05/08/2015 - See CHANGELOG @ 201505061810
 # 06/20/2015 - See CHANGELOG @ 201506200420
 # 07/05/2015 - See CHANGELOG @ 201506290420
+# 07/11/2015 - See CHANGELOG @ 201507110420
 # ---------------------------------------------------------------------------------------------------|
 thisFile=$(echo "$0" | sed 's/\.\///g')
 echo "${thisFile}" | awk '{print toupper($0)}'
@@ -31,7 +32,8 @@ then
     echo "[t] - type. Valid types include: "
     echo ""
     echo -e "\t\"dojo\"\t\tDojo Fiddle"
-    echo -e "\t\"extjs\"\t\tExt JS Fiddle"
+    echo -e "\t\"extjs5\"\t\tExt JS 5 Fiddle"
+    echo -e "\t\"extjs6\"\t\tExt JS 6 Fiddle"
     echo -e "\t\"php\"\t\tPHP Fiddle"
     echo -e "\t\"jquery\"\tjQuery / Bootstrap Fiddle"
     echo -e "\t\"three\"\t\tthree.js / WebGl Fiddle"
@@ -57,7 +59,7 @@ echo ${bornOnDate}
 #try
 (
     case ${type} in
-        'extjs'|'php'|'jquery'|'three'|'dojo'|'node'|'tween'|'svg')
+        'extjs5'|'extjs6'|'php'|'jquery'|'three'|'dojo'|'node'|'tween'|'svg')
             case ${type} in
                 'php') fiddleName=$(echo "$fiddleNameStub.php";)
                     ;;