+++ scripts/sencha-build-testing.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-cd ~/github/house/web/house
-sencha app build testing
-