+++ scripts/sencha-enable-dc.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-sudo chmod u+w ../web/house/bootstrap.js
-