+++ scripts/sencha-examples-start.sh
@@ -1,3 +1,0 @@
-cd ~/github/house/web/ext/build/examples
-sencha web start
-