+++ scripts/sencha-start-testing.sh
@@ -1,3 +1,0 @@
-cd ~/github/house/web/build/testing/House
-sencha web start
-