+++ scripts/sencha-theme-build.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-cd ~/github/house/web/packages/house-theme
-sencha package build
-